refactor(artifact): decompose the cursor engine into send + paint (phase 6a)
use-live-cursors.ts was a 448-line hook doing two distinct jobs. Split it along
the natural seam:
- use-cursor-send.ts — our own pointer → server (throttled) + the focus/idle/leave
  state machine.
- use-cursor-paint.ts — peer frames → the rAF-driven overlay (ease/lerp, name-tag
  fade, off-screen edge indicators, roster/ripple state).
use-live-cursors.ts stays the public entry (35 lines): it mints the shared per-tab
`selfId`, composes the two halves, and re-exports the overlay types so cursor-layer
imports are unchanged. selfId is the only shared value — send stamps it on outgoing
frames, paint ignores our own echoed frames; each half subscribes to the cursor pref
independently (pure useContext).

Behavior-preserving — line-for-line faithful. Verified against the UNCHANGED original:
under --repeat-each stress the one timing-sensitive capture-side test flakes in BOTH
(original 3/4, split 2/4), so the brittleness is pre-existing, not introduced. Under
the normal config all cursor + deck e2e pass; full gate green.
